hw-jpa: домашка по Spring Data JPA

+++ books/src/main/java/com/books/books/service/AuthorServiceImpl.java
@@ -2,9 +2,9 @@ package com.books.books.service;
 
 import com.books.books.models.Author;
 import com.books.books.models.Book;
-import com.books.books.models.Style;
-import com.books.books.repositories.AuthorRepositoryJpa;
-import com.books.books.repositories.BookRepositoryJpa;
+import com.books.books.repositoriesSpringDataJPA.AuthorRepository;
+import com.books.books.repositoriesSpringDataJPA.BookRepository;
+import com.books.books.repositoriesSpringDataJPA.StyleRepository;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.stereotype.Service;
@@ -12,7 +12,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Scanner;
@@ -24,13 +23,18 @@ public class AuthorServiceImpl implements AuthorService {
     @PersistenceContext
     EntityManager em;
 
-    private final AuthorRepositoryJpa authorRepository;
-    private final BookRepositoryJpa bookRepository;
+    private final AuthorRepository authorRepository;
+    private final BookRepository bookRepository;
+    private final StyleRepository styleRepository;
     private final BookService bookService;
 
-    public AuthorServiceImpl(AuthorRepositoryJpa authorRepository, BookRepositoryJpa bookRepository, BookService bookService) {
+    public AuthorServiceImpl(AuthorRepository authorRepository,
+                             BookRepository bookRepository,
+                             StyleRepository styleRepository,
+                             BookService bookService) {
         this.authorRepository = authorRepository;
         this.bookRepository = bookRepository;
+        this.styleRepository = styleRepository;
         this.bookService = bookService;
     }
 
@@ -38,7 +42,7 @@ public class AuthorServiceImpl implements AuthorService {
     @Transactional
     @ShellMethod(value = "Print all authors", key = {"a1"})
     public void printAuthors() {
-        List<Author> authors = authorRepository.findAll();
+        List<Author> authors = (List<Author>) authorRepository.findAll();
         for (Author author : authors) {
             String authorText = " ID: " + author.getId() + "; \n Автор: " + author.getAuthorName();
             System.out.println(authorText);
@@ -78,19 +82,17 @@ public class AuthorServiceImpl implements AuthorService {
             System.out.println("Имя автора не может быть пустым");
             return;
         }
-        if (!Objects.equals(authorRepository.findByName(authorName), null)) {
+        if (!Objects.equals(authorRepository.findByAuthorNameContains(authorName), null)) {
             System.out.println("Такой автор уже добавлен");
             return;
         }
         Author author = new Author();
         author.setAuthorName(authorName);
-        long authorId = authorRepository.save(author).orElse(0L);
-        if (authorId != 0) {
-            Author newAuthor = authorRepository.findById(authorId).orElse(new Author());
-            System.out.println("Добавлен автор: \n" +
-                    " ID: " + newAuthor.getId() + "; \n Автор: " + newAuthor.getAuthorName());
-        }
+        authorRepository.save(author);
+        System.out.println("Добавлен автор: \n" +
+                " ID: " + author.getId() + "; \n Автор: " + author.getAuthorName());
     }
+
 
     @Override
     @Transactional
@@ -137,7 +139,12 @@ public class AuthorServiceImpl implements AuthorService {
         }
         Author author = authorRepository.findById(authorId).orElse(new Author());
         if (!(author.getId() == 0)) {
-            authorRepository.deleteById(authorId);
+            if(author.getAuthorBooks().isEmpty()) {
+                authorRepository.deleteById(authorId);
+            }
+            else{
+                System.out.println("Удалить автора невозможно - сначала удалите все книги этого автора");
+            }
         } else {
             System.out.println("Не найдено автора по id: " + authorId);
         }
